Creating Subscriber for the card issuance queue

+++ mscreditanalyzer/src/main/java/io/github/cursomservice/mscreditanalyzer/domain/model/ClientData.java
@@ -3,10 +3,12 @@ package io.github.cursomservice.mscreditanalyzer.domain.model;
 import lombok.Data;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Data
 @Getter
 public class ClientData {
     private Long id;
     private String name;
-    private Integer age;
+    private BigDecimal age;
 }
